requirements.txt
asgiref==3.3.1
astroid==2.4.2
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dadata==20.7.0
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dash-ui==0.4.0
Django==3.1.4
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
GDAL==3.2.0
gunicorn==20.0.4
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
idna==2.10
image==1.5.33
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mercantile==1.1.6
numpy==1.19.4
osgeo==0.0.0
osm2geojson==0.1.28
pandas==1.1.5
Pillow==8.0.1
plotly==4.14.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
retrying==1.3.3
rfc3986==1.4.0
Shapely==1.7.1
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
